test: address diff update review comments
